% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{add_node}
\alias{add_node}
\title{Add a `Node` to a `Tree`}
\usage{
add_node(x, ...)
}
\arguments{
\item{x}{Some `Tree`.}

\item{...}{Further arguments - unused at present.}
}
\description{
Add a `Node` to a `Tree`
}
