% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{add_node.Tree}
\alias{add_node.Tree}
\title{Add a `Node` to a `Tree`}
\usage{
\method{add_node}{Tree}(x, node, ...)
}
\arguments{
\item{x}{Some `Tree`.}

\item{node}{Some `Node` to be added to the `Tree`.}

\item{...}{Further arguments - unused at present.}
}
\description{
Add a `Node` to a `Tree`
}
