% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{as_node}
\alias{as_node}
\title{Convert a `list` into a `Node`}
\usage{
as_node(x)
}
\arguments{
\item{x}{A list of data. This must contain entries named
`name`, `parent` and `children`.}
}
\description{
Convert a `list` into a `Node`
}
