% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{children.Node}
\alias{children.Node}
\title{Extract the names of the children of a `Node`}
\usage{
\method{children}{Node}(x, ...)
}
\arguments{
\item{x}{A `Node`}

\item{...}{Other arguments - unused at present.}
}
\description{
Extract the names of the children of a `Node`
}
