% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{get_node.Tree}
\alias{get_node.Tree}
\title{Get a `Node` using it's name from a `Tree`}
\usage{
\method{get_node}{Tree}(x, node_name, ...)
}
\arguments{
\item{x}{A `Tree`.}

\item{node_name}{The name of a `Node` within the `Tree`.}

\item{...}{Other arguments: unused at present.}
}
\description{
Get a `Node` using it's name from a `Tree`
}
