% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{get_parent}
\alias{get_parent}
\title{Obtain the `Node` that is the parent of a given `Node` in the `Tree`}
\usage{
get_parent(x, ...)
}
\arguments{
\item{x}{A `Tree`.}

\item{...}{Other arguments.}
}
\description{
Obtain the `Node` that is the parent of a given `Node` in the `Tree`
}
