% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{get_parent.Tree}
\alias{get_parent.Tree}
\title{Obtain the `Node` that is the parent of a given `Node` in the `Tree`}
\usage{
\method{get_parent}{Tree}(x, node, ...)
}
\arguments{
\item{x}{A `Tree`}

\item{node}{A `Node` that has a parent within the `Tree`.}

\item{...}{Other arguments - currently unused.}
}
\description{
Obtain the `Node` that is the parent of a given `Node` in the `Tree`
}
