% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{has_parent}
\alias{has_parent}
\title{Does this `Node` have a parent in the `Tree`?}
\usage{
has_parent(x, ...)
}
\arguments{
\item{x}{a `Node`}

\item{...}{further args - not currently used.}
}
\description{
Does this `Node` have a parent in the `Tree`?
}
