% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{is_empty}
\alias{is_empty}
\title{Returns true for Trees that don't contain `Node`s}
\usage{
is_empty(x, ...)
}
\arguments{
\item{x}{A `Tree` object.}

\item{...}{Other arguments - unused at present.}
}
\description{
Returns true for Trees that don't contain `Node`s
}
