% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{node}
\alias{node}
\title{Construct a `Node`}
\usage{
node(name = "root", parent = NULL, children = NULL, ...)
}
\arguments{
\item{name}{Name of the `Node`}

\item{parent}{Name of the (lone) parent of the `Node`}

\item{children}{Names of any children of the `Node`}

\item{...}{Any other fields of data for storing with the
`Node`. Provided as key=value pairs.}
}
\description{
Construct a `Node`
}
