% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{nodes}
\alias{nodes}
\title{Obtain the `Node`s that are stored in a `Tree`}
\usage{
nodes(x, ...)
}
\arguments{
\item{x}{A `Tree` object.}

\item{...}{Other arguments - unused at present.}
}
\description{
Obtain the `Node`s that are stored in a `Tree`
}
