% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{nodes.Tree}
\alias{nodes.Tree}
\title{Obtain the `Node`s that are stored in a `Tree`}
\usage{
\method{nodes}{Tree}(x, ...)
}
\arguments{
\item{x}{A `Tree` object.}

\item{...}{Other arguments - unused at present.}
}
\description{
Obtain the `Node`s that are stored in a `Tree`
}
