% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{parent_name.Node}
\alias{parent_name.Node}
\title{parent_name of a node}
\usage{
\method{parent_name}{Node}(x, ...)
}
\arguments{
\item{x}{a node.}

\item{...}{further args - not currently used.}
}
\description{
parent_name of a node
}
