% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{update_node}
\alias{update_node}
\title{Update the data that is attached to a `Node`}
\usage{
update_node(x, ...)
}
\arguments{
\item{x}{a `Node`}

\item{...}{a set of key=value pairs for appending as fields
to the current `Node`.}
}
\description{
Update the data that is attached to a `Node`
}
