% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{map_tree}
\alias{map_tree}
\title{Map a function over the `Node`s in a `Tree`}
\usage{
map_tree(.x, .f, ..., .field = NULL)
}
\arguments{
\item{.x}{The `Tree`}

\item{.f}{The function to be applied to the `Node`s in the
`Tree`. If .f has an argument called `tree` then `map_tree` will pass the
`Tree` `.x` in as that argument automatically.}

\item{...}{Further arguments to be passed into the function
`.f`. Except: if the function `.f` takes an argument `tree`, then `.x` is
passed in as the `tree` argument automatically.}

\item{.field}{If a string is provided, the values of `.f` will
be appended to the respective `Node`s in the `Tree`; .field gives the name
of the field where this result is stored. If .field is NULL (as default)
a named list of the results will be returned (names being the name of the
relevant `Node`).}
}
\description{
Map a function over the `Node`s in a `Tree`
}
